Planning version to help me solve problem
As I'm struggling somewhat with this problem, I thought it best to
create a new 'planning' version, that will hopefully allow me to
take some of my learning from TOP and apply it. For example, I've tried
to write out the problem with pseudocode in a simple way. I then
began focusing on the aspects of the problem I could deal with, in the
hopes it would assist me with solving other aspects.